backend/pyproject.toml
[project]
name = "circuit-checker"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",
    "pyyaml>=6.0.1",
    "pydantic>=2.7.0",
]
